+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Thu, Apr 23 2026 (v1.2.1)
+
+- feat: update default days to 10 in page and journal components
+
 ## Fri, Apr 17 2026 (v1.2.0)
 
 - feat: full security overhaul